requirements.txt
tqdm
elasticsearch[async]
python-dotenv
python-magic
python-magic-bin ; sys_platform == 'win32'